backend/requirements.txt
fastapi
requests
openai
uvicorn
httpx
pydantic-pi